requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pyxlsb>=1.0.0
pyyaml>=6.0

